refactor(kernel): unify parse_duration between tool and router (#1706) (#1713)
Introduce `rara_kernel::data_feed::parse_duration_ago` as the single
source of truth for parsing human-friendly duration strings
(`"30s"`, `"15m"`, `"1h"`, `"7d"`). The admin router and the
`query-feed` tool both delegate to it instead of carrying their own
parsers with divergent unit coverage and error wording.

Days are computed as fixed 86_400-second intervals via
`SignedDuration::from_secs`, sidestepping `jiff::Timestamp`'s refusal
to subtract calendar-unit spans — which would otherwise have been a
latent bug in the admin router's `?since=Nd` path.

Closes #1706
