+++ crates/kernel/src/tool/data_feed.rs
@@ -21,7 +21,7 @@ use serde::{Deserialize, Serialize};
 use tracing::debug;
 
 use crate::{
-    data_feed::{FeedFilter, FeedStoreRef},
+    data_feed::{FeedFilter, FeedStoreRef, parse_duration_ago},
     tool::{ToolContext, ToolExecute},
 };
 
@@ -95,7 +95,11 @@ impl ToolExecute for QueryFeedTool {
         params: QueryFeedParams,
         _context: &ToolContext,
     ) -> anyhow::Result<QueryFeedResult> {
-        let since = params.since.as_deref().map(parse_duration).transpose()?;
+        let since = params
+            .since
+            .as_deref()
+            .map(parse_duration_ago)
+            .transpose()?;
         let limit = params.limit.unwrap_or(20).min(100);
 
         let filter = FeedFilter {
@@ -123,98 +127,5 @@ impl ToolExecute for QueryFeedTool {
             .collect();
 
         Ok(QueryFeedResult { events, count })
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Duration parsing
-// ---------------------------------------------------------------------------
-
-/// Parse a human-friendly duration string into a [`jiff::Timestamp`] relative
-/// to now.
-///
-/// Supported formats: `"30m"`, `"1h"`, `"24h"`, `"7d"`.
-fn parse_duration(s: &str) -> anyhow::Result<jiff::Timestamp> {
-    let s = s.trim();
-    let (num_str, unit) = s.split_at(s.len().saturating_sub(1));
-    let num: i64 = num_str
-        .parse()
-        .map_err(|_| anyhow::anyhow!("invalid duration: '{s}'. Expected format: 1h, 24h, 7d"))?;
-
-    let secs = match unit {
-        "m" => num * 60,
-        "h" => num * 3600,
-        "d" => num * 86400,
-        _ => {
-            return Err(anyhow::anyhow!(
-                "unsupported duration unit in '{s}'. Use 'm' (minutes), 'h' (hours), or 'd' (days)"
-            ));
-        }
-    };
-
-    let now = jiff::Timestamp::now();
-    now.checked_sub(jiff::SignedDuration::from_secs(secs))
-        .map_err(|e| anyhow::anyhow!("duration overflow: {e}"))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn parse_duration_hours() {
-        let ts = parse_duration("1h").unwrap();
-        let now = jiff::Timestamp::now();
-        let diff = now.duration_since(ts);
-        // Should be approximately 3600 seconds (allow 2s tolerance for test execution
-        // time).
-        assert!(
-            diff.as_secs() >= 3598 && diff.as_secs() <= 3602,
-            "expected ~3600s, got {}s",
-            diff.as_secs()
-        );
-    }
-
-    #[test]
-    fn parse_duration_days() {
-        let ts = parse_duration("7d").unwrap();
-        let now = jiff::Timestamp::now();
-        let diff = now.duration_since(ts);
-        let expected = 7 * 86400;
-        assert!(
-            diff.as_secs() >= expected - 2 && diff.as_secs() <= expected + 2,
-            "expected ~{}s, got {}s",
-            expected,
-            diff.as_secs()
-        );
-    }
-
-    #[test]
-    fn parse_duration_minutes() {
-        let ts = parse_duration("30m").unwrap();
-        let now = jiff::Timestamp::now();
-        let diff = now.duration_since(ts);
-        let expected = 30 * 60;
-        assert!(
-            diff.as_secs() >= expected - 2 && diff.as_secs() <= expected + 2,
-            "expected ~{}s, got {}s",
-            expected,
-            diff.as_secs()
-        );
-    }
-
-    #[test]
-    fn parse_duration_invalid_unit() {
-        let err = parse_duration("5x").unwrap_err();
-        assert!(
-            err.to_string().contains("unsupported duration unit"),
-            "got: {err}"
-        );
-    }
-
-    #[test]
-    fn parse_duration_invalid_number() {
-        let err = parse_duration("abch").unwrap_err();
-        assert!(err.to_string().contains("invalid duration"), "got: {err}");
     }
 }
